Split gulpfile into smaller task files
