+++ tasks/gulp/lint.js
@@ -1,0 +1,28 @@
+'use strict'
+
+const gulp = require('gulp')
+const paths = require('../../config/paths.json')
+const standard = require('gulp-standard')
+const sasslint = require('gulp-sass-lint')
+
+// Javascript lint check -----------------
+// ---------------------------------------
+gulp.task('js:lint', () => {
+  return gulp.src([paths.components + '**/*.js'])
+    .pipe(standard())
+    .pipe(standard.reporter('default', {
+      breakOnError: true,
+      quiet: true
+    }))
+})
+
+// Scss lint check -----------------------
+// ---------------------------------------
+gulp.task('scss:lint', () => {
+  return gulp.src(paths.src + '**/*.scss')
+    .pipe(sasslint({
+      configFile: paths.config + '.sass-lint.yml'
+    }))
+    .pipe(sasslint.format())
+    .pipe(sasslint.failOnError())
+})
